slideshow: introduce fakeRequestAnimationFrame for hidden tab mode
When we hide main tab with COOL in the browser - slideshow is not
refreshed at all or that often as usual.
In this commit we simulate requestAnimationFrame by manual
callback invoke at around 60 fps to refresh the canvas.
As can be seen in Chrome-like browsers - real refresh happens
with 1 fps or when user event happens.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I69989e7a63e2681e0dc96a4302e5740f7849d8e9