fix(calc): Improve zoom view-jumping
Previously, we got significant view jumps in calc by virtue of an
incorrect center position. To fix this, I've melded together the old
formula (from before commit d6f375c501690370e41ceeec18200599bd33490c)
with the new formula. I've gated this to calc as the old formula is
calc specific. More testing will be needed to determine if there is any
jumping in writer/impress.

There are still minor jumps from the end of the animation to the final
zoom pos, which is particularly noticable in split view, but I think
this may actually be *better* than the animated-to position, so instead
of further iterating on this, I'd like to look and see if we should
change that instead.

This code still doesn't affect tablet view.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I7127d8f0a3156ed9dfb04bdd5fb801c318531269